+++ .eslintignore
@@ -1,4 +1,4 @@
-.*
+!.*
 .git/
 .github/
 coverage/
