+++ .storybook/manager.js
@@ -1,5 +1,5 @@
 import { addons } from '@storybook/addons'
 
 addons.setConfig({
-  panelPosition: 'right'
+  panelPosition: 'right',
 })
