+++ package.json
@@ -66,25 +66,35 @@
   ],
   "scripts": {
     "build": "rm -rf ./dist && rollup -c rollup.config.js && cp ./source/styles.css ./dist/styles.css",
-    "build:docs": "rm -rf ./docs && mkdir ./docs && build-storybook -o docs",
-    "ci": "concurrently npm:lint npm:test npm:build npm:build:docs",
+    "build:docs": "rm -rf ./docs && mkdir ./docs && build-storybook --quiet -o docs",
+    "ci": "concurrently npm:lint npm:build npm:build:docs",
     "contributors:add": "all-contributors add",
     "contributors:generate": "all-contributors generate",
     "lint": "eslint .",
-    "prepublishOnly": "concurrently npm:lint npm:test npm:build",
+    "prepublishOnly": "concurrently npm:lint npm:build",
     "start": "start-storybook -p 6006",
-    "test": "jest"
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook",
+    "test-storybook": "test-storybook"
   },
   "devDependencies": {
+    "@babel/core": "^7.18.9",
+    "@mdx-js/react": "^2.1.2",
     "@rollup/plugin-typescript": "^8.3.3",
     "@storybook/addon-a11y": "^6.5.9",
-    "@storybook/addon-controls": "^6.5.9",
-    "@storybook/react": "^6.5.9",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^14.3.0",
+    "@storybook/addon-actions": "^6.5.10-alpha.0",
+    "@storybook/addon-docs": "^6.5.9",
+    "@storybook/addon-interactions": "^6.5.10-alpha.0",
+    "@storybook/addon-links": "^6.5.10-alpha.0",
+    "@storybook/addon-viewport": "^6.5.9",
+    "@storybook/builder-webpack4": "^6.5.10-alpha.0",
+    "@storybook/jest": "^0.0.10",
+    "@storybook/manager-webpack4": "^6.5.10-alpha.0",
+    "@storybook/mdx2-csf": "^0.0.3",
+    "@storybook/react": "^6.5.10-alpha.0",
+    "@storybook/test-runner": "^0.6.0--canary.154.5e2f62a.0",
+    "@storybook/testing-library": "^0.0.13",
     "@types/axe-core": "^2.0.7",
-    "@types/jest": "^28.1.6",
     "@types/node": "^18.6.3",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
@@ -92,24 +102,23 @@
     "@typescript-eslint/parser": "^5.31.0",
     "all-contributors-cli": "^6.20.0",
     "axe-core": "^4.4.3",
+    "babel-loader": "^8.2.5",
     "concurrently": "^7.3.0",
     "eslint": "^8.20.0",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "^26.7.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-n": "^15.2.4",
     "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-react": "^7.30.1",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-storybook": "^0.6.1",
     "jest": "^28.1.3",
-    "jest-environment-jsdom": "^28.1.3",
     "npm-run-all": "^4.1.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "rollup": "^2.77.2",
     "rollup-plugin-dts": "^4.2.2",
-    "ts-jest": "^28.0.7",
     "typescript": "^4.7.4"
   },
   "peerDependencies": {
