--- source/Uncontrolled.test.tsx
@@ -1,3 +1,0 @@
-test('truthy', async () => {
-  expect(true).toEqual(true)
-})
