+++ stories/Introduction.stories.mdx
@@ -1,0 +1,16 @@
+import { Meta } from '@storybook/addon-docs'
+import './base.css'
+
+<Meta title="Introduction" />
+
+# Introduction
+
+This site showcases different usages of
+[`react-medium-image-zoom`](https://github.com/rpearce/react-medium-image-zoom),
+the original medium.com-inspired image zooming library for React.
+
+Use the links in the sidebar to try out different features that the library
+supports.
+
+For more info, please consult the [`react-medium-image-zoom`
+README](https://github.com/rpearce/react-medium-image-zoom#readme).
